requirements.txt
PyGithub==1.43.6
pyramid==1.10.3
firebase-admin==2.16.0
google-cloud-firestore==0.32.1
cryptography==2.6.1
python-box
